requirements.txt
openai
langchain_nvidia_ai_endpoints
langchain_community
faiss-cpu
python-dotenv
streamlit
pypdf
# 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
# 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